+++ hash.h
@@ -77,16 +77,16 @@ hash_set_stmt (struct hash *hash,
 int 
 hash_set_param_count (struct hash *hash,
          uint32_t laddr, uint32_t raddr, uint16_t lport, uint16_t rport,
-         int stmt_id, int param_count);
+         ulong stmt_id, int param_count);
 
 int
 hash_set_param (struct hash *hash, 
-         uint32_t laddr, uint32_t raddr, uint16_t lport, uint16_t rport,struct timeval value, int stmt_id,
+         uint32_t laddr, uint32_t raddr, uint16_t lport, uint16_t rport,struct timeval value, ulong stmt_id,
           char *param, char *param_type, int param_count);
 
 int
 hash_get_param_count(struct hash *hash,
          uint32_t laddr, uint32_t raddr, uint16_t lport, uint16_t rport,
-         int stmt_id, int *param_count, char **param_type);
+         ulong stmt_id, int *param_count, char **param_type);
 
 #endif
